requirements.txt
absl-py==2.2.1
accelerate==1.5.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
albucore==0.0.24
albumentations==2.0.8
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrdict==2.0.1
attrs==25.3.0
binpacking==1.5.2
bitsandbytes==0.45.4
bleach==6.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.5
contourpy==1.3.1
cpm-kernels==1.0.11
crcmod==1.7
cryptography==44.0.2
cycler==0.12.1
dacite==1.9.2
datasets==3.2.0
deepspeed==0.16.5
# Editable install with no version control (diffusers==0.33.0.dev0)
-e /home/xinran/xinran/generation/janus_model/Janus-main/diffusers
dill==0.3.8
distro==1.9.0
einops==0.8.1
et_xmlfile==2.0.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
gradio==3.48.0
gradio_client==0.6.1
grpcio==1.71.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
# Editable install with no version control (janus==1.0.0)
-e /home/xinran/xinran/generation/janus_model/Janus-main
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==0.10.0
joblib @ file:///croot/joblib_1718217211762/work
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kaggle==1.7.4.2
kiwisolver==1.4.8
latex2mathml==3.77.0
lazy_loader==0.4
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.1
mdtex2html==1.3.0
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
modelscope==1.24.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.2.0
multiprocess==0.70.16
narwhals==1.32.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.4.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.69.0
opencv-python==4.11.0.86
opencv-python-headless==4.13.0.92
openpyxl==3.1.5
orjson==3.10.16
oss2==2.19.1
packaging==24.2
pandas==2.2.3
peft==0.15.1
pillow==10.4.0
prefetch_generator==1.0.3
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.0
pydantic_core==2.33.0
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.3
pypinyin==0.50.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge==1.0.1
rpds-py==0.24.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn @ file:///croot/scikit-learn_1737988764076/work
scipy @ file:///croot/scipy_1743153247252/work/dist/scipy-1.15.2-cp312-cp312-linux_x86_64.whl#sha256=de6f647f5e7e8f92c700731aa3da0614b12980ac8670287948695cf9ffd072f7
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==69.5.1
simplejson==3.20.1
simsimd==6.5.16
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
starlette==0.46.1
stringzilla==4.6.0
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
text-unidecode==1.3
thop==0.1.1.post2209072238
threadpoolctl @ file:///croot/threadpoolctl_1719407800858/work
tifffile==2025.6.11
tiktoken==0.5.2
timm==1.0.15
tokenizers==0.21.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.50.2
transformers-stream-generator==0.0.5
triton==3.2.0
trl==0.16.0
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.45.1
xformer==1.0.1
xformers==0.0.29.post3
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
